* 修复下载路径不存在时，程序关闭不了的问题 * 双击视频某p Title可获取作品信息+avId(适用于批量打开的情况) * 长按视频某p Title可更新预览图(适用于批量打开的情况) * 点击预览图可获取图片链接 * `parser`遍历选择时增加break，减少不必要的循环 * 解决文件路径中含多个空格问题

+++ src/nicelee/bilibili/util/CmdUtil.java
@@ -202,10 +202,12 @@ public class CmdUtil {
 				return false;
 			}
 		};
-		// 删除下载文件
-		for (File file : folderDown.listFiles(filter)) {
-			System.out.println("尝试删除" + file.getName());
-			file.delete();
+		if(folderDown.exists()) {
+			// 删除下载文件
+			for (File file : folderDown.listFiles(filter)) {
+				System.out.println("尝试删除" + file.getName());
+				file.delete();
+			}
 		}
 	}
